@@ -967,7 +967,7 @@ namespace TVRename
 
         private void PopulateShowStatusColours([NotNull] TVSettings s)
         {
-            if (s.ShowStatusColors == null)
+            if (s.ShowStatusColors is null)
             {
                 return;
             }
@@ -1267,7 +1267,7 @@ namespace TVRename
                     if (!ColorTranslator.FromHtml(txtShowStatusColor.Text).IsEmpty && ssct != null)
                     {
                         ListViewItem item = lvwDefinedColors.FindItemWithText(ssct.Text);
-                        if (item == null)
+                        if (item is null)
                         {
                             item = new ListViewItem();
                             item.SubItems.Add(txtShowStatusColor.Text);
