@@ -235,12 +235,12 @@ namespace TVRename
         {
             ActionMede8erXML nfo = o as ActionMede8erXML;
 
-            if (Episode == null)
+            if (Episode is null)
             {
                 return 1;
             }
 
-            if (nfo?.Episode == null)
+            if (nfo?.Episode is null)
             {
                 return -1;
             }
