@@ -110,12 +110,12 @@ namespace TVRename
         {
             ActionPyTivoMeta nfo = o as ActionPyTivoMeta;
 
-            if (Episode == null)
+            if (Episode is null)
             {
                 return 1;
             }
 
-            if (nfo?.Episode == null)
+            if (nfo?.Episode is null)
             {
                 return -1;
             }
