@@ -11,38 +11,63 @@ using System.Windows.Forms;
 
 namespace TVRename
 {
-    public sealed class DateSorterWtw : ListViewItemSorter
+    public abstract class ListViewItemDateSorter : ListViewItemSorter
     {
-        public DateSorterWtw(int column) : base(column) { }
+        public ListViewItemDateSorter(int column) : base(column) { }
 
         protected override int CompareListViewItem(ListViewItem x, ListViewItem y)
         {
             DateTime? d1 = GetDate(x);
             DateTime? d2 = GetDate(y);
 
-            if ((d1 == null) && (d2 == null))
+            if ((d1 is null) && (d2 is null))
             {
                 return 0;
             }
 
-            if (d1 == null)
+            if (d1 is null)
             {
                 return -1;
             }
 
-            if (d2 == null)
+            if (d2 is null)
             {
                 return 1;
             }
 
             return d1.Value.CompareTo(d2.Value);
         }
+        protected abstract DateTime? GetDate(ListViewItem lvi);
+    }
 
-        private static DateTime? GetDate(ListViewItem lvi)
+    public class DateSorterWtw : ListViewItemDateSorter
+    {
+        public DateSorterWtw(int column) : base(column) { }
+
+        protected override DateTime? GetDate(ListViewItem lvi)
         {
             try
             {
                 return ((ProcessedEpisode)(lvi.Tag)).GetAirDateDt(true);
+            }
+            catch
+            {
+                return DateTime.Now;
+            }
+        }
+    }
+
+    public class DateSorterScan : ListViewItemDateSorter
+    {
+        public DateSorterScan(int column) : base(column) { }
+        protected override DateTime? GetDate(ListViewItem lvi)
+        {
+            try
+            {
+                ProcessedEpisode e = ((Item)(lvi.Tag)).Episode;
+                if (e is null) return DateTime.Now;
+
+                return e.GetAirDateDt(true);
             }
             catch
             {
