@@ -228,7 +228,7 @@
             this.Name = "ChooseFile";
             this.ShowIcon = false;
             this.ShowInTaskbar = false;
-            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterParent;
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
             this.Text = "Choose File";
             this.TopMost = true;
             this.ResumeLayout(false);
