@@ -6,6 +6,8 @@
 // Copyright (c) TV Rename. This code is released under GPLv3 https://github.com/TV-Rename/tvrename/blob/master/LICENSE.md
 // 
 
+using JetBrains.Annotations;
+
 namespace TVRename
 {
     public class SeriesSpecifier
@@ -15,12 +17,21 @@ namespace TVRename
         public readonly string CustomLanguageCode;
         public readonly string Name;
 
-        public SeriesSpecifier(int key, bool useCustomLanguage, string customLanguageCode,string name)
+        public SeriesSpecifier(int key, bool useCustomLanguage, [CanBeNull] string customLanguageCode,string name)
         {
             SeriesId = key;
-            UseCustomLanguage = useCustomLanguage;
-            CustomLanguageCode = customLanguageCode;
             Name = name;
+
+            if (string.IsNullOrWhiteSpace(customLanguageCode))
+            {
+                UseCustomLanguage = false;
+                CustomLanguageCode = TVSettings.Instance.PreferredLanguageCode;
+            }
+            else
+            {
+                UseCustomLanguage = useCustomLanguage;
+                CustomLanguageCode = customLanguageCode;
+            }
         }
     }
 }
